Indelwald 1.0 - Part I
This involves all functions required to stringently filter and analyse
indel outputs from Pindel and Platypus. A toy example with 3 anonymised
hosts and tumours was created. In order to go through the functions in
a step-by-step manner, just open “1_Indelwald_main.R”, and follow the
flow.

Note, that 5 R-packages need to be installed on your machine to get all
functions to run properly: VennDiagram, gridExtra, stringr, data.table,
and GenomicRanges.

Enjoy Indelwald!